requirements.txt
yfinance
pandas
praw
requests
pyyaml
matplotlib
openai
numpy
Flask